database tables modified

+++ public/js/admin/products/delete.js
@@ -7,53 +7,59 @@ document.addEventListener('DOMContentLoaded', function() {
         buttons[i].addEventListener('click', deleteProduct);
 
     }
+
+    function deleteProduct() {
+
+        var id = this.dataset.id;
+    
+        if (confirm('Are you sure?')) {
+            
+            if (id != null && id != '') {
+    
+                let data = { id: id };
+    
+                fetch('/products/delete', {
+    
+                    method: "DELETE",
+                    headers: { 'Content-Type': 'application/json' },
+                    body: JSON.stringify(data)
+    
+                })
+                .then(r => {
+    
+                    return r.json();
+    
+                })
+                .then(r => {
+    
+                    if (r.ok) {
+    
+                        this.parentNode.parentNode.previousElementSibling.innerHTML = `<span class="alert alert-success">${r.msg}</span>`;
+                        let that = this;
+                        setTimeout(function(){
+
+                            that.parentNode.parentNode.parentNode.remove();
+
+                        }, 3000);
+    
+                    }
+                    else {
+    
+                        alert(r.msg);
+    
+                    }
+    
+                })
+                .catch(e => {
+    
+                    console.log(e);
+    
+                });
+    
+            }
+            
+        }
+    
+    }
     
 });
-
-function deleteProduct() {
-
-    var id = this.dataset.id;
-
-    if (confirm('Are you sure?')) {
-        
-        if (id != '') {
-
-            let data = { id: id };
-
-            fetch('/products/delete', {
-
-                method: "DELETE",
-                headers: { 'Content-Type': 'application/json' },
-                body: JSON.stringify(data)
-
-            })
-            .then(r=> {
-
-                return r.json();
-
-            })
-            .then(r => {
-
-                if (r.ok) {
-
-                    window.location.reload();
-
-                }
-                else {
-
-                    alert('Error on delete product');
-
-                }
-
-            })
-            .catch(e => {
-
-                console.log(e);
-
-            });
-
-        }
-        
-    }
-
-}
